Clean up Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
