+++ implementation/src/main/java/io/smallrye/graphql/index/Annotations.java
@@ -26,9 +26,9 @@ import org.eclipse.microprofile.graphql.Id;
 import org.eclipse.microprofile.graphql.Ignore;
 import org.eclipse.microprofile.graphql.InputType;
 import org.eclipse.microprofile.graphql.Mutation;
+import org.eclipse.microprofile.graphql.Name;
 import org.eclipse.microprofile.graphql.NonNull;
 import org.eclipse.microprofile.graphql.Query;
-import org.eclipse.microprofile.graphql.SchemaName;
 import org.eclipse.microprofile.graphql.Type;
 import org.jboss.jandex.DotName;
 
@@ -58,5 +58,5 @@ public interface Annotations {
     public static final DotName BEAN_VALIDATION_NOT_EMPTY = DotName.createSimple("javax.validation.constraints.NotEmpty");
     public static final DotName BEAN_VALIDATION_NOT_BLANK = DotName.createSimple("javax.validation.constraints.NotBlank");
 
-    public static final DotName SCHEMANAME = DotName.createSimple(SchemaName.class.getName());
+    public static final DotName NAME = DotName.createSimple(Name.class.getName());
 }
